requirements.txt
streamlit
requests
pandas
plotly
psycopg2-binary